+++ src/test/java/org/springframework/data/couchbase/transactions/CouchbaseReactiveTransactionsWrapperTemplateIntegrationTests.java
@@ -59,7 +59,7 @@ import com.couchbase.client.java.transactions.error.TransactionFailedException;
  * Tests for ReactiveTransactionsWrapper, using template methods (findById etc.)
  */
 @IgnoreWhen(missesCapabilities = Capabilities.QUERY, clusterTypes = ClusterType.MOCKED)
-@SpringJUnitConfig(TransactionsConfigCouchbaseSimpleTransactionManager.class)
+@SpringJUnitConfig(TransactionsConfig.class)
 public class CouchbaseReactiveTransactionsWrapperTemplateIntegrationTests extends JavaIntegrationTests {
 	// intellij flags "Could not autowire" when config classes are specified with classes={...}. But they are populated.
 	@Autowired ReactiveCouchbaseClientFactory couchbaseClientFactory;
