+++ src/test/java/org/springframework/data/couchbase/transactions/CouchbaseTransactionsWrapperTemplateIntegrationTests.java
@@ -59,7 +59,7 @@ import com.couchbase.client.java.transactions.error.TransactionFailedException;
  * Tests for TransactionsWrapper, using template methods (findById etc.)
  */
 @IgnoreWhen(missesCapabilities = Capabilities.QUERY, clusterTypes = ClusterType.MOCKED)
-@SpringJUnitConfig(TransactionsConfigCouchbaseSimpleTransactionManager.class)
+@SpringJUnitConfig(TransactionsConfig.class)
 public class CouchbaseTransactionsWrapperTemplateIntegrationTests extends JavaIntegrationTests {
 	// intellij flags "Could not autowire" when config classes are specified with classes={...}. But they are populated.
 	@Autowired CouchbaseClientFactory couchbaseClientFactory;
