--- src/test/java/org/springframework/data/couchbase/transactions/TransactionsConfigCouchbaseSimpleTransactionManager.java
@@ -1,60 +1,0 @@
-package org.springframework.data.couchbase.transactions;
-
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Role;
-import org.springframework.data.couchbase.config.AbstractCouchbaseConfiguration;
-import org.springframework.data.couchbase.repository.config.EnableCouchbaseRepositories;
-import org.springframework.data.couchbase.repository.config.EnableReactiveCouchbaseRepositories;
-import org.springframework.data.couchbase.transaction.CouchbaseSimpleTransactionInterceptor;
-import org.springframework.data.couchbase.util.ClusterAwareIntegrationTests;
-import org.springframework.transaction.TransactionManager;
-import org.springframework.transaction.annotation.AnnotationTransactionAttributeSource;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import org.springframework.transaction.interceptor.TransactionAttributeSource;
-import org.springframework.transaction.interceptor.TransactionInterceptor;
-
-/**
- * See comments on TransactionsConfigCouchbaseTransactionManager for why this test exists.
- */
-@Configuration
-@EnableCouchbaseRepositories("org.springframework.data.couchbase")
-@EnableReactiveCouchbaseRepositories("org.springframework.data.couchbase")
-@EnableTransactionManagement
-public class TransactionsConfigCouchbaseSimpleTransactionManager extends AbstractCouchbaseConfiguration {
-
-	@Override
-	public String getConnectionString() {
-		return ClusterAwareIntegrationTests.connectionString();
-	}
-
-	@Override
-	public String getUserName() {
-		return ClusterAwareIntegrationTests.config().adminUsername();
-	}
-
-	@Override
-	public String getPassword() {
-		return ClusterAwareIntegrationTests.config().adminPassword();
-	}
-
-	@Override
-	public String getBucketName() {
-		return ClusterAwareIntegrationTests.bucketName();
-	}
-
-	@Bean
-	@Role(BeanDefinition.ROLE_INFRASTRUCTURE)
-	public TransactionInterceptor transactionInterceptor(TransactionManager couchbaseTransactionManager) {
-		TransactionAttributeSource transactionAttributeSource = new AnnotationTransactionAttributeSource();
-		TransactionInterceptor interceptor = new CouchbaseSimpleTransactionInterceptor(couchbaseTransactionManager, transactionAttributeSource);
-		interceptor.setTransactionAttributeSource(transactionAttributeSource);
-		if (couchbaseTransactionManager != null) {
-			interceptor.setTransactionManager(couchbaseTransactionManager);
-		}
-		return interceptor;
-	}
-
-}
